webhook-server/requirements.txt
# Webhook Dashboard Service Requirements

# FastAPI and WebSocket
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Notifications
slack-sdk>=3.27.0
aiohttp>=3.9.0
aiosmtplib>=3.0.0
boto3>=1.29.0  # Optional: for SES API mode

# Database
aiosqlite>=0.19.0  # Optional: for SQLite persistence

# Utilities
tenacity>=8.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Notion integration
notion-client>=2.2.0

# Google Workspace Events API Integration (optional)
google-cloud-pubsub>=2.18.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
cloudevents>=1.10.0
